[#158] refactor: 타입분리 및 로직 분리
