+++ src/app/(with-header)/myactivity/components/CategoryInput.tsx
@@ -1,9 +1,5 @@
 import ChevronIcon from '@assets/svg/chevron'; // 아이콘 경로는 맞게 조정
-
-interface CategoryProps {
-  category?: string;
-  onCategoryChange: (value: string) => void;
-}
+import { CategoryProps } from '@/types/addEditExperienceType';
 
 export default function CategoryInput({
   category,
@@ -16,7 +12,7 @@ export default function CategoryInput({
       </label> */}
       <div className='relative w-full'>
         <select
-          className={`appearance-none w-full rounded-md border border-gray-800 bg-white px-20 py-17 ${
+          className={`w-full appearance-none rounded-md border border-gray-800 bg-white px-20 py-17 ${
             category ? 'text-black' : 'text-gray-400'
           }`}
           id='category'
@@ -26,16 +22,28 @@ export default function CategoryInput({
           <option value='' disabled hidden>
             카테고리
           </option>
-          <option className='text-black' value='문화 · 예술'>문화 · 예술</option>
-          <option className='text-black' value='식음료'>식음료</option>
-          <option className='text-black' value='스포츠'>스포츠</option>
-          <option className='text-black' value='투어'>투어</option>
-          <option className='text-black' value='관광'>관광</option>
-          <option className='text-black' value='웰빙'>웰빙</option>
+          <option className='text-black' value='문화 · 예술'>
+            문화 · 예술
+          </option>
+          <option className='text-black' value='식음료'>
+            식음료
+          </option>
+          <option className='text-black' value='스포츠'>
+            스포츠
+          </option>
+          <option className='text-black' value='투어'>
+            투어
+          </option>
+          <option className='text-black' value='관광'>
+            관광
+          </option>
+          <option className='text-black' value='웰빙'>
+            웰빙
+          </option>
         </select>
 
         {/* 커스텀 화살표 아이콘 */}
-        <div className='pointer-events-none absolute right-12 top-1/2 -translate-y-1/2'>
+        <div className='pointer-events-none absolute top-1/2 right-12 -translate-y-1/2'>
           <ChevronIcon size={20} />
         </div>
       </div>
