+++ src/types/bookingInterfaceType.ts
@@ -1,0 +1,7 @@
+export interface BookingButtonProps {
+  onClick: () => void;
+  children: React.ReactNode;
+  disabled?: boolean;
+  onBooking?: boolean;
+  className?: string;
+}
